requirements.txt
torch
transformers
accelerate
sentence-transformers
faiss-cpu
PyMuPDF
numpy
streamlit